add new and old callbacks
